Add friend data in today fragment
