--- model/src/main/java/jt/projects/model/Friend.kt
@@ -1,13 +1,0 @@
-package jt.projects.model
-
-import java.time.LocalDate
-
-data class Friend(
-    val idVk: Long?,
-    val idPhone: String?,
-    val type: FriendType,
-    val name: String,
-    val birthDate: LocalDate,
-    val age: Int,
-    val photoUrl: String
-)
